fix imports and some refactor
