+++ osrs-wiki-scraper/bis_graph/bis_graph.py
@@ -1,16 +1,18 @@
 from dataclasses import dataclass
 from typing import List
 
-from bis_constants import Style, STYLE_STATS
-from bis_item import BisItem, BisItemWalker
-from bis_visual_graph import BisVisualGraph
-from generate_web_app_data import GenerateWebAppData
-from wiki_data import WikiData
+from bis_graph.bis_constants import Style, STYLE_STATS
+from bis_graph.bis_item import BisItem, BisItemWalker
+from bis_graph.bis_visual_graph import BisVisualGraph
+from bis_graph.wiki_data import WikiData
+from util import get_attack_style_and_type, is_filtered_item
 
 # twisted, blorva, other stuff
 FILTER_IDS = [
     "24664", "24666", "24668", "28254", "13199", "13197", "28256", "28258",
-    "28687", "28688", "28682"
+    "28687", "28688", "28682", "26695", "26721", "26722", "26695",
+    "11705", "21276", "11706",
+    "27374", "27376", "24780"
 ]
 
 
@@ -21,7 +23,7 @@ class BisItemsGraph:
 
 class GenerateBisItems:
     def __init__(self):
-        WikiData.load_all()  # TODO kinda scuffed but to have global static wiki data
+        WikiData.load_all()
 
     def create_bis_items(self):
         seen_item_names = []
@@ -37,16 +39,13 @@ class GenerateBisItems:
         })
 
         for item_id, item in WikiData.items.items():
-            if item_id == "12608":  # TODO remove
-                print(item_id)
-                a = 1
-            if GenerateWebAppData.is_filtered_item(item, item_id):
+            if is_filtered_item(item, item_id):
                 continue
 
             if item["name"] in seen_item_names:
                 continue
 
-            if self.is_filtered_item(item, item_id):
+            if self.is_bis_filtered_item(item, item_id):
                 continue
 
             seen_item_names.append(item["name"])
@@ -57,17 +56,12 @@ class GenerateBisItems:
 
             slot = item["slot"]
             for style in styles:
-                if style != Style.MELEE or slot != 5:
-                    continue  # TODO remove
-
                 if slot not in bis_item_root.items[style]:
                     new_bis_item = BisItem([item_id], [], [])
                     bis_item_root.items[style][slot] = [new_bis_item]
                     bis_item_leaf.items[style][slot] = [new_bis_item]
                     continue
 
-                # root_bis_items = BisItem([], [], bis_item_graph.items[style][slot])
-                # leaf_bis_items = BisItem([], bis_item_leaf.items[style][slot], [])
                 BisItemWalker.insert_item(
                     bis_item_root.items[style][slot], bis_item_leaf.items[style][slot], style, item_id
                 )
@@ -75,7 +69,7 @@ class GenerateBisItems:
         visual = BisVisualGraph()
         visual.create_graph_image(bis_item_root)
 
-    def is_filtered_item(self, item, item_id):
+    def is_bis_filtered_item(self, item, item_id):
         if any(i == item_id for i in FILTER_IDS):
             return True
         if "(perfected)" in item["name"]:
@@ -115,7 +109,7 @@ class GenerateBisItems:
 
     def get_item_style(self, item) -> list[Style]:
         styles = []
-        attack_styles, attack_type = GenerateWebAppData.get_attack_style_and_type(item)
+        attack_styles, attack_type = get_attack_style_and_type(item)
         if attack_type:
             if item.get("str", 0) < 0 or item.get("rstr", 0) < 0 or item.get("mdmg", 0) < 0:
                 return []
@@ -137,7 +131,7 @@ class GenerateBisItems:
             elif style in styles:
                 styles.remove(style)
 
-        return styles
+        return list(set(styles))
 
 
 if __name__ == '__main__':
