+++ osrs-wiki-scraper/bis_graph/bis_visual_graph.py
@@ -6,9 +6,9 @@ from typing import List
 from PIL import Image
 from graphviz import Digraph
 
-from bis_constants import Style
-from wiki_data import WikiData
-from bis_item import BisItem
+from bis_graph.bis_constants import Style
+from bis_graph.wiki_data import WikiData
+from bis_graph.bis_item import BisItem
 
 
 class BisVisualGraph:
@@ -17,17 +17,11 @@ class BisVisualGraph:
         height = 5000
         final_image = Image.new('RGBA', (width, height))
 
-        x_offset = 0
-        y_offset = 0
-        x_diff = 40
-        y_diff = 40
-        for style in [Style.MELEE, Style.RANGED, Style.MAGIC]:  # TODO other styles
+        for style in [Style.MELEE, Style.RANGED, Style.MAGIC]:
             for slot in bis_graph.items[style]:
                 print("slot: " + str(slot))
-                items = WikiData.gear_slot_items[str(slot)]
-                # bis_items = bis_graph.items[style][slot]
                 self.create_spring_graph_image(bis_graph.items[style][slot],
-                                               "graphs/" + str(style.name) + "_slot_" + str(slot))
+                                               "bis_graph/graphs/" + str(style.name) + "_slot_" + str(slot))
         return final_image
 
     def create_spring_graph_image(self, items: List[BisItem], output_path: str):
@@ -35,9 +29,6 @@ class BisVisualGraph:
         dot = Digraph()
 
         img_dir = 'img'
-
-        # Dictionary to hold item_id to BisItem mappings
-        # item_dict = {item.item_id: item for item in items} # TODO remove?
 
         # Set to keep track of visited nodes
         visited = set()
@@ -54,18 +45,18 @@ class BisVisualGraph:
             for item_id in item.item_ids:
                 icon_base64 = self.get_icon(item_id)
                 if icon_base64 is None:
-                    print("skip: " + str(item_id))
+                    print("error getting icon: " + str(item_id))
                     return
 
                 try:
                     icon_image = Image.open(io.BytesIO(base64.b64decode(icon_base64)))
                 except Exception:
-                    print("skip: " + str(item_id))
+                    print("error decoding icon: " + str(item_id))
                     return
 
                 icon_path = os.path.join(img_dir, f'{item_id}.png')
                 icon_paths.append(icon_path)
-                icon_image.save(os.path.join("graphs", icon_path))
+                icon_image.save(os.path.join("bis_graph", "graphs", icon_path))
 
                 label_html += '<td>' + WikiData.items[item_id]["name"] + "(" + str(item_id) + ")" + '</td>'
                 icons_html += '<td><img src="' + icon_path + '"/></td>'
